correcion

+++ budget-optimizer-backend/src/main/java/com/budgetoptimizer/budget_optimizer_backend/dto/ml/AnalisisPresupuestoResponse.java
@@ -14,5 +14,5 @@ public class AnalisisPresupuestoResponse {
     private String analisis;
     private List<String> recomendaciones;
     private String ahorroPotencial;
-    private Double confianza;
+    private BigDecimal confianza;
 }
