+++ budget-optimizer-backend/src/main/java/com/budgetoptimizer/budget_optimizer_backend/dto/ml/Anomalia.java
@@ -14,6 +14,6 @@ public class Anomalia {
     private String categoria;
     private BigDecimal montoEsperado;
     private BigDecimal montoReal;
-    private Double desviacion;
+    private BigDecimal desviacion;
     private String descripcion;
 }
